+++ src/PlutoVIP/VR_PhoneInstalledBinary.h
@@ -25,14 +25,14 @@
 class VR_PhoneInstalledBinary : public RA_Request
 {
 	// Request Variables
-	u_int64_t m_iMacAddress;
+	unsigned long long m_iMacAddress;
 	unsigned long m_iRevision;
 
 	// Response Variables
 
 public:
 	// The establishment will call this constructor, then ConvertRequestToBinary
-	VR_PhoneInstalledBinary(u_int64_t MacAddress,unsigned long Revision);
+	VR_PhoneInstalledBinary(unsigned long long MacAddress,unsigned long Revision);
 	VR_PhoneInstalledBinary() {}
 
 	virtual unsigned long ID() { return VRS_PHONE_INSTALLED_BIN; }
